requirements-dev.txt
coverage
funcparserlib==0.3.6
mockldap==0.2.6
pycodestyle==2.9.1
pyflakes==2.5.0
flake8==5.0.4

pytest==7.0.1
py==1.11.0
pytest-cov==3.0.0
pytest-django==4.5.2